finish initial implementation of the create prompt command, better organize file structure
